refactor(coordinator): 把 #[cfg(test)] mod tests 移到 coordinator/tests.rs
机械式行为保持：把 ~850 行单元测试模块体搬到独立文件，父模块改为 #[cfg(test)] mod tests;
声明。tests 里的 super::* / super::dictation::* 从子文件解析路径一致，417 测试不变。
